backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
greenlet>=3.0
aiosqlite>=0.20
pydantic>=2.10
pydantic-settings>=2.7
python-dotenv>=1.0
google-genai>=1.0
httpx>=0.27
numpy>=1.26
